Use render prop function pattern
The function receives as arguments the cursor props, which can then be
passed into the relevant downstream nodes.

This approach makes it straightforward to implement TypeScript
definitions and is in line with the approach taken by many react-based
libraries.

In fact, the react documentation on render props uses a mouse tracking
component as its prototypical example: https://reactjs.org/docs/render-props.html